refactor(caveSearch): add CaveAutoCompleteSearchComponent
